Add lame, naive hot loop benchmark to do a quick compare with a similarly naive Criterion benchmark at https://github.com/luketpeterson/fuzzy_rocks to get this Nim impl being about 1.5x faster on hot loop queries than the Rust/C++ RocksDB version (at least on the one machine I tested it on for the tokyo/london examples).  To be fair, I only do ASCII not utf8 words.
In any event, it would be better to do re-do in Rust the more scientific
"suggest for batches of misspellings in some document" model otherwise
analyzed here.  However, that is much more work than these 2 lines of
code (and more Rust than I probably ever want to personally deal with).